+++ src/simulation/simulation_impl.rs
@@ -124,7 +124,7 @@ impl SimulationImpl {
             location: physics_body.location,
             rotation: physics_body.rotation,
             mobility: physics_body.mobility,
-            passable: self.world.is_body_passable(body_handle),
+            passable: physics_body.passable,
             associated_data: non_physical_object_data.associated_data.clone(),
         })
     }
@@ -544,9 +544,6 @@ mod tests {
         world
             .expect_body(partial_eq(returned_handle))
             .returns(Some(expected_physical_body));
-        world
-            .expect_is_body_passable(partial_eq(returned_handle))
-            .returns(expected_passable);
 
         let mut simulation = SimulationImpl::new(
             world,
@@ -596,9 +593,6 @@ mod tests {
         world
             .expect_body(partial_eq(returned_handle))
             .returns(Some(expected_physical_body));
-        world
-            .expect_is_body_passable(partial_eq(returned_handle))
-            .returns(expected_passable);
 
         let mut simulation = SimulationImpl::new(
             world,
@@ -948,9 +942,6 @@ mod tests {
         world
             .expect_body(partial_eq(returned_handle))
             .returns(Some(expected_physical_body.clone()));
-        world
-            .expect_is_body_passable(partial_eq(returned_handle))
-            .returns(expected_physical_body.passable);
 
         let mut simulation = SimulationImpl::new(
             box world,
@@ -1016,9 +1007,6 @@ mod tests {
         world
             .expect_body(partial_eq(returned_handle))
             .returns(Some(expected_physical_body.clone()));
-        world
-            .expect_is_body_passable(partial_eq(returned_handle))
-            .returns(expected_physical_body.passable);
 
         let mut simulation = SimulationImpl::new(
             box world,
@@ -1053,9 +1041,6 @@ mod tests {
         world
             .expect_body(partial_eq(returned_handle))
             .returns(Some(expected_physical_body.clone()));
-        world
-            .expect_is_body_passable(partial_eq(returned_handle))
-            .returns(expected_physical_body.passable);
 
         let mut simulation = SimulationImpl::new(
             box world,